+++ src/VennDiagram/VennArc.tsx
@@ -114,17 +114,17 @@ export const VennArc: FC<Partial<VennArcProps>> = ({
   stroke,
   mask,
   id,
-  active = false,
   style,
-  inactiveStyle = { opacity: 0.3 },
-  activeStyle = { opacity: 0.8 },
-  initialStyle = { opacity: 0.6 },
-  strokeWidth = 3,
-  gradient = <Gradient />,
-  tooltip = <ChartTooltip />,
-  onClick = () => undefined,
-  onMouseEnter = () => undefined,
-  onMouseLeave = () => undefined
+  active,
+  inactiveStyle,
+  activeStyle,
+  initialStyle,
+  strokeWidth,
+  gradient,
+  tooltip,
+  onClick,
+  onMouseEnter,
+  onMouseLeave
 }) => {
   const [internalActive, setInternalActive] = useState<boolean>(false);
   const arcRef = useRef<any | null>(null);
@@ -211,4 +211,17 @@ export const VennArc: FC<Partial<VennArcProps>> = ({
       )}
     </g>
   );
+};
+
+VennArc.defaultProps = {
+  active: false,
+  inactiveStyle: { opacity: 0.3 },
+  activeStyle: { opacity: 0.8 },
+  initialStyle: { opacity: 0.6 },
+  strokeWidth: 3,
+  gradient: <Gradient />,
+  tooltip: <ChartTooltip />,
+  onClick: () => undefined,
+  onMouseEnter: () => undefined,
+  onMouseLeave: () => undefined
 };