+++ src/VennDiagram/VennSeries.tsx
@@ -70,12 +70,12 @@ export const VennSeries: FC<Partial<VennSeriesProps>> = ({
   data,
   id,
   selections,
-  animated = true,
-  disabled = false,
-  colorScheme = 'cybertron',
-  outerLabel = <VennOuterLabel />,
-  arc = <VennArc />,
-  label = <VennLabel />
+  animated,
+  disabled,
+  colorScheme,
+  outerLabel,
+  arc,
+  label
 }) => {
   const transition = animated ? DEFAULT_TRANSITION : { type: false, delay: 0 };
   const [actives, setActives] = useState<string[]>([]);
@@ -226,4 +226,13 @@ export const VennSeries: FC<Partial<VennSeriesProps>> = ({
       ))}
     </motion.g>
   );
+};
+
+VennSeries.defaultProps = {
+  animated: true,
+  disabled: false,
+  colorScheme: 'cybertron',
+  outerLabel: <VennOuterLabel />,
+  arc: <VennArc />,
+  label: <VennLabel />
 };