+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Master
 - [chore] update deps
-- [chore] migrate piechart to fc
+- [chore] migrate Pie Chart to fc
+- [chore] migrate Venn Diagram default props
 
 # 11.1.1- 9/28/21
 - [improvement] add color inversion to Venn and Bubble charts